Added the first test for ReviewReader.
Doing so helps me to consider how I'd like the reader to work and behave when used in the app.

I ended up adding a constructor for testing purposes - clearly named as such. Perhaps there's a better way of designing the tests and/or the class being tested - thoughts welcome.