A basic results screen added to show some information about the file the user loaded.
This needs significant enhancements. I'll probably move much of the functionality to the newly added Activity, which will also make it easier to provide better information to the user, and allow them to load files. It might be a good time to add some duplicate file checking, to reduce the chances of users repeatedly loading the same file.