Added simple code to detect the format of the CSV file picked by a user.
This code is mainly to help improve the information we can show users if they try to load a file that has the wrong content. The Google Play Developer Console files are probably consistent in how they're formatted (as UTF-16 with the BOM provided). I've tested the new code with various CSV files, but yet to find one that includes the BOM for UTF-8. I've not written tests for the new code, instread I've spent a while using the debugger and trying to learn about how the various files were encoded and how to write suitable code in Java.

Thanks to various StackOverflow posts for ideas, and to
http://www.ericbt.com/Blog/60 and
http://unicodebook.readthedocs.io/guess_encoding.html which I may yet implement.

I'm closing https://github.com/julianharty/app-store-reviews-app/issues/11 now this has been implemented.