Experimenting in-situ to improve the parsing of CSV data with commas in the contents.
I'll move the method to the app's codebase next.

Note: the calls to split now include -1 as the second parameter so that lines are
fully split even when there are empty columns.

Thanks to various StackOverflow questions and answers which helped me learn more about how to solve the challenge of parsing CSV data more completely.