Finally, the first version with the RecyclerView for the Triage process.
The Triage process doesn't work yet. I need to add the capability. I also need to consider how to interact with the database correctly so we don't lose memory or have performance issues. And the GUI is very unattractive. Nonetheless this is a significant step in the right direction, finally. Phew!