Here's a version of the app with Google Firebase added.
I only did a minimal integration to see what was involved. I'm likely to remove Firebase soon (later today) and use Google Analytics instead. Google Analytics has a much larger body of information available and it's - dare I say it - a more traditional in-app mobile analytics service. Firebase is broader and more encompasing. It may well be a better choice as I gain experience and as my research and the app both mature.

For some reason Android Studio and/or the Firebase automated integration updated the target Java from Java 7 to Java 8. Again I may revert this as it wasn't intentional.