Well, a couple of hours later I've switched from Firebase to Google Analytics.
It's not particularly clear how to setup Google analytics to report specific data. I hope I'll discover ways to work effectively with it soon. I can now see the data is being reported by Google Analytics but not see the data in the Online Console. There are some other minor edits in this commit and I noticed we're back to Java 7.